+++ VsIntegration/Spect.Net.VsPackage/CustomEditors/AsmEditor/Z80AsmTokenTagger.cs
@@ -84,6 +84,8 @@ namespace Spect.Net.VsPackage.CustomEditors.AsmEditor
                 var currentLine = curSpan.Start.GetContainingLine();
                 var textOfLine = currentLine.GetText();
 
+                // TODO: Implement coloring of the block comment (/* ... */)
+
                 // --- Let's use the Z80 assembly parser to obtain tags
                 var inputStream = new AntlrInputStream(textOfLine);
                 var lexer = new Z80AsmLexer(inputStream);